netns: limit parallelism to $(nproc) tests at once
The parallel tests were added to catch queueing issues from multiple
cores. But what happens in reality when testing tons of processes is
that these separate threads wind up fighting with the scheduler, and we
wind up with contention in places we don't care about that decrease the
chances of hitting a bug. So just do a test with the number of CPU
cores, rather than trying to scale up arbitrarily.

Change-Id: I0220869592fa89a8f7fdafd1ac4a42b81779c346
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>